% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_search.R
\name{global_search}
\alias{global_search}
\title{General Step in The Global Search Stage of TikTak}
\usage{
global_search(
  i,
  s,
  visited_minimum,
  local_search_fun,
  theta_min = 0.1,
  theta_max = 0.995,
  theta_pow = 0.5
)
}
\arguments{
\item{i}{the \eqn{i}th step}

\item{s}{the matrix of candidate points/parameters from the initial step}

\item{visited_minimum}{best result from previous local searches}

\item{local_search_fun}{function perform local search}

\item{theta_min}{minimum weight}

\item{theta_max}{maximum weight}

\item{theta_pow}{determines the rate of increase in weight}
}
\value{
The best result from local searches.
}
\description{
This function implements the ith step the global search stage of the TikTak
algorithm.
}
\details{
\code{visited_minimum} must be a list with the same structure as the
results from \code{nloptr}. Similarly, \code{local_search_fun} is function of
a single (vector) parameter that produces a list with the same structure as
\code{visited_minimum}.
}
